refactor(clients): extract shared KaguraRestClient base (#229)
Files/Resource/Secret/Workspace clients carried four near-verbatim
copies of the construction/auth/lifecycle/error spine, already drifted
(403 scrubbing and the OAuth-aware 401 hint existed in some, not
others — #228 review). _rest_base.KaguraRestClient now owns:

- credential requirement (ValueError naming the concrete class),
  HTTPS validation, static-Bearer vs OAuth httpx client construction
- the from_mcp_url/_from_resolved_auth factory pair (#115 pairing)
- async lifecycle and the _request spine with _raise_401/403/429
  hooks; _json/_expect_list body guards

Per-client remainders: Files keeps the unauth R2 upload leg (dual
close) and the workspace-403 hint hook; Resource keeps _mcp_url
stamping; Secret keeps the existence-hiding 403 and its historical
generic 429; Workspace keeps the owner-key 403 hint and
detail-carrying 429.

Zero public-API change; all 1526 tests green unchanged except one
monkeypatch target updated (workspace from_mcp_url test now patches
_rest_base._resolve_auth — the function moved). Deliberate
unifications: FilesClient's 401 hint is now OAuth-aware like its
siblings, and Retry-After parsing uses the tolerant
_retry_after_seconds helper instead of bare int(). Net -713/+118
across the four clients; _rest_base is 100% covered by the existing
per-client suites.
